deployment-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyyaml==6.0.1
docker==6.1.3
aiohttp==3.9.1
pydantic==2.5.0
python-multipart==0.0.6
